LayoutTests/canvas/philip/tests/2d.imageData.get.length-expected.txt

< [index] >
2d.imageData.get.length
getImageData() returns a correctly-sized CanvasPixelArray
References: 2d.pixelarray.length
Actual output:
Passed
